_archive/projects/fullpotential_ai/fullpotential_core/droplets/hteam/droplet-2/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
psutil==5.9.6
python-jose[cryptography]==3.3.0
PyJWT==2.10.1
httpx==0.25.1
structlog==23.2.0
python-dotenv==1.0.0
requests==2.31.0